fix: update code for breaking changes in Python 3.11
Fixes #127. Struct members were hidden from PyFrameObject and PyThreadState. See https://docs.python.org/3/whatsnew/3.11.html#pyframeobject-3-11-hiding. This PR is passing public and internal tests.

However, I am not particularly confident that this fix is safe/correct. The handler operates possibly without the GIL, as it does not currently modify any of the PyObjects it works with. It gets the thread local ThreadState with `PyGILState_GetThisThreadState()`.
- `PyCodeObject`s may be GC'ed while the handler is running, but we get around that by patching the PyCode deallocator function [`PyCode_Type.tp_dealloc = &CodeDealloc`](https://github.com/GoogleCloudPlatform/cloud-profiler-python/blob/ae0c31b7ed962c8158e8a24ba44be6442c7374f8/googlecloudprofiler/src/profiler.h#L60)

The problem is that the getters described in [the changelog](https://docs.python.org/3/whatsnew/3.11.html#pyframeobject-3-11-hiding) (`PyThreadState_GetFrame()`, `PyFrame_GetCode()`, `PyFrame_GetBack()`) increment the refcount for those PyObjects which afaict is NOT safe to do without holding the GIL. Trying to acquire the GIL in the handler seems to deadlock some of the internal integration tests; I'm also not confident that this would be [async-signal-safe](https://man7.org/linux/man-pages/man7/signal-safety.7.html).

Folks are welcome to try the profiler in Python 3.11 from this PR but I am not comfortable merging it as is. If you do try it, please report back on if you encountered any issues.

PiperOrigin-RevId: 499370977